# backend/swagger/swagger.yml
openapi: 3.0.3
info:
  title: Task Management API
  description: REST API for the Task Management application.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tasks:
    get:
      tags: [task]
      summary: List tasks
      operationId: listTasks

      parameters:
        - name: limit
          in: query

          schema:
            type: integer
            default: 20

        - name: offset
          in: query

          schema:
            type: integer
            default: 0

      responses:
        "200":
          description: Successful operation

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [task]
      summary: Create task
      operationId: createTask

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"

      responses:
        "201":
          description: Task created

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "400":
          description: Bad request

          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/users:
    post:
      tags: [user]
      summary: Create user
      operationId: createUser

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"

      responses:
        "201":
          description: User created

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        "400":
          description: Bad request

          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"

        "409":
          $ref: "#/components/responses/DuplicateEmailError"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  ##############################
  # Reusable responses
  ##############################
  responses:
...
    UnauthorizedError:
      description: Unauthorized

      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error

      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    DuplicateEmailError:
      description: Duplicate email

      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object

      properties:
        code:
          type: string

        message:
          type: string

      required:
        - code
        - message
    ValidationErrorResponse:
      type: object

      properties:
        code:
          type: string

        message:
          type: string

        errors:
          type: object

          additionalProperties:
            type: string

      required:
        - code
        - message
        - errors

    ##############################
    # ENUM
    ##############################
    Status:
      type: string

      enum:
        - TODO
        - DOING
        - DONE

    ##############################
    # Task
    ##############################
    Task:
      type: object

      properties:
        id:
          type: integer
          format: int64

        user_id:
          type: integer
          format: int64

        title:
          type: string

        description:
          type: string

        status:
          $ref: "#/components/schemas/Status"

        due_date:
          type: string
          format: date-time
          nullable: true

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time
    TaskWithUser:
      type: object

      properties:
        task_id:
          type: integer
          format: int64

        title:
          type: string

        status:
          $ref: "#/components/schemas/Status"

        user_id:
          type: integer
          format: int64

        user_email:
          type: string
    TaskListResponse:
      type: object

      properties:
        count:
          type: integer

        items:
          type: array

          items:
            $ref: "#/components/schemas/TaskWithUser"

        limit:
          type: integer

        offset:
          type: integer
    CreateTaskRequest:
      type: object

      required:
        - user_id
        - title
        - status

      properties:
        user_id:
          type: integer
          format: int64

        title:
          type: string

        description:
          type: string

        status:
          $ref: "#/components/schemas/Status"

        due_date:
          type: string
          format: date-time
...
    User:
      type: object

      properties:
        id:
          type: integer
          format: int64

        email:
          type: string

        display_name:
          type: string

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object

      required:
        - email
        - display_name

      properties:
        email:
          type: string

        display_name:
          type: string
...